refactor(index): simplify option parsing
Previously, option parsing was insane. `owner` and `repo` could be set
either independently or in a string, path was optional, and overall it
felt very hacky. With this commit, API has to deal with a string form of
`"owner/repo[/path]"` and an `options` object, or either one of these.
Also, it would be easy now to extend it to get a `branch` argument (#2).

BREAKING CHANGE: callsites using separate `owner`, `repo`, or `path`
arguments must be changed.